fix: Catch and wrap InstantiationException
Catch and wrap java.lang.InstantiationException if it is thrown when
we try to instantiate a class via reflection. The caught exception is
then put inside a DependecyException which is then rethrown.

The purpose of this change is to be able to provide more context on
which dependency we tried to resolve and build when this exception is
thrown.